fix code smell

+++ multi_physics/QED/include/picsar_qed/containers/picsar_tables.hpp
@@ -872,7 +872,7 @@ namespace containers{
                 throw std::runtime_error("raw_data contains invalid data.");
             if(m_how_many_y <= 0)
                 throw std::runtime_error("raw_data contains invalid data.");
-            m_values = VectorType(m_how_many_x*m_how_many_y);
+            m_values = VectorType(static_cast<int>(m_how_many_x*m_how_many_y));
             auto vals = serialization::get_n_out<RealType>(
                     it_raw_data,
                     m_how_many_x*m_how_many_y);
@@ -1067,7 +1067,7 @@ namespace containers{
         std::vector<std::array<RealType,2>> get_all_coordinates() const noexcept
         {
             auto all_coords = std::vector<std::array<RealType,2>>(
-                m_how_many_x*m_how_many_y, {0,0});
+                static_cast<int>(m_how_many_x*m_how_many_y), {0,0});
             int count = 0;
             for (int i = 0; i < m_how_many_x; ++i){
                 for (int j = 0; j < m_how_many_y; ++j){
